fix: repair pytest collection and bound dbt below 2.0.0
The test suite did not run at all. pytest 9 (pinned via pytest>=9.0.1)
removed support for defining `pytest_plugins` in a non-root conftest, so
`uv run pytest .` failed during collection with zero tests executed:

    Defining 'pytest_plugins' in a non-top-level conftest is no longer
    supported: /unit_tests/conftest.py

Move unit_tests/conftest.py to the repo root, which is where pytest now
requires it. This also makes `from definitions import get_test_configs`
resolve deliberately rather than by accident -- the root conftest is what
places the repo root on sys.path. 13 tests now collect.

While moving it, drop the GITHUB_ACTIONS branch and its service-account
keyfile path. `method: oauth` resolves through Google Application Default
Credentials, which covers both local development (gcloud auth
application-default login) and CI (Workload Identity Federation exports
GOOGLE_APPLICATION_CREDENTIALS). One code path, no static key, and no
CI-specific logic living in test code.

Bound dbt-core and dbt-bigquery below 2.0.0. Both floors were unbounded,
and dbt-core 2.0.0b1 published on 2026-08-10; resolvers skip prereleases
by default, but an unbounded floor pulls 2.0.0 in silently the day it
goes final. Resolution is unchanged (dbt-core 1.10.15, dbt-bigquery
1.10.3) -- the ceiling is a guard, not a bump.

Add .env.example, since the suite requires BIGQUERY_PROJECT but shipped
no template, and track package-lock.yml so a fresh clone resolves the
pinned dbt_utils 1.3.2 rather than a floating version.
