+++ .env.example
@@ -1,0 +1,16 @@
+# Copy this file to `.env` at the repo root (it is gitignored) and fill in the value
+# below. `.env` is loaded automatically by pytest-dotenv when you run the test suite.
+
+# The GCP project that the unit tests build their throwaway datasets in.
+# Each test class creates and drops its own `test<timestamp>_<module>` dataset here,
+# so this must be a sandbox project you are comfortable writing to.
+BIGQUERY_PROJECT=
+
+# Authentication is not configured here. The tests use dbt-bigquery's `method: oauth`,
+# which resolves through Google Application Default Credentials. Log in once with:
+#
+#   gcloud auth application-default login \
+#     --scopes=https://www.googleapis.com/auth/bigquery,https://www.googleapis.com/auth/iam.test
+#
+# In CI, Workload Identity Federation exports GOOGLE_APPLICATION_CREDENTIALS and ADC
+# picks it up, so the same code path is used locally and in CI.
