+++ package-lock.yml
@@ -1,0 +1,5 @@
+packages:
+  - name: dbt_utils
+    package: dbt-labs/dbt_utils
+    version: 1.3.2
+sha1_hash: dd1e1feb2d2bbce79e7a255cd309a60e6548df0b
