+++ pyproject.toml
@@ -5,8 +5,8 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
-    "dbt-bigquery>=1.10.3",
-    "dbt-core>=1.10.15",
+    "dbt-bigquery>=1.10.3,<2.0.0",
+    "dbt-core>=1.10.15,<2.0.0",
 ]
 
 [dependency-groups]
